Move get resource functions to kube_helpers, move simple helper functions to util.go
