+++ pkg/kubernetes/kube_helper.go
@@ -1,8 +1,9 @@
-package common
+package kube
 
 import (
 	"context"
 
+	"github.com/keikoproj/kubedog/pkg/common"
 	"github.com/pkg/errors"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
@@ -10,9 +11,11 @@ import (
 	"k8s.io/client-go/kubernetes"
 )
 
+var ()
+
 // ListPodsWithLabelSelector lists pods with a label selector
 func ListPodsWithLabelSelector(client kubernetes.Interface, namespace, selector string) (*corev1.PodList, error) {
-	pods, err := RetryOnError(&DefaultRetry, IsRetriable, func() (interface{}, error) {
+	pods, err := common.RetryOnError(&common.DefaultRetry, common.IsRetriable, func() (interface{}, error) {
 		return client.CoreV1().Pods(namespace).List(context.Background(), metav1.ListOptions{LabelSelector: selector})
 	})
 	if err != nil {
@@ -24,7 +27,7 @@ func ListPodsWithLabelSelector(client kubernetes.Interface, namespace, selector 
 
 // ListNodes lists nodes
 func ListNodes(client kubernetes.Interface) (*corev1.NodeList, error) {
-	nodes, err := RetryOnError(&DefaultRetry, IsRetriable, func() (interface{}, error) {
+	nodes, err := common.RetryOnError(&common.DefaultRetry, common.IsRetriable, func() (interface{}, error) {
 		return client.CoreV1().Nodes().List(context.Background(), metav1.ListOptions{})
 	})
 	if err != nil {
@@ -36,7 +39,7 @@ func ListNodes(client kubernetes.Interface) (*corev1.NodeList, error) {
 
 // GetDaemonset gets a daemonset
 func GetDaemonset(client kubernetes.Interface, name, namespace string) (*appsv1.DaemonSet, error) {
-	ds, err := RetryOnError(&DefaultRetry, IsRetriable, func() (interface{}, error) {
+	ds, err := common.RetryOnError(&common.DefaultRetry, common.IsRetriable, func() (interface{}, error) {
 		return client.AppsV1().DaemonSets(namespace).Get(context.Background(), name, metav1.GetOptions{})
 	})
 	if err != nil {
@@ -47,7 +50,7 @@ func GetDaemonset(client kubernetes.Interface, name, namespace string) (*appsv1.
 
 // GetDeployment gets a deployment
 func GetDeployment(client kubernetes.Interface, name, namespace string) (*appsv1.Deployment, error) {
-	deploy, err := RetryOnError(&DefaultRetry, IsRetriable, func() (interface{}, error) {
+	deploy, err := common.RetryOnError(&common.DefaultRetry, common.IsRetriable, func() (interface{}, error) {
 		return client.AppsV1().Deployments(namespace).Get(context.Background(), name, metav1.GetOptions{})
 	})
 	if err != nil {
@@ -58,7 +61,7 @@ func GetDeployment(client kubernetes.Interface, name, namespace string) (*appsv1
 
 // GetPersistentVolume gets a pv
 func GetPersistentVolume(client kubernetes.Interface, name string) (*corev1.PersistentVolume, error) {
-	pvs, err := RetryOnError(&DefaultRetry, IsRetriable, func() (interface{}, error) {
+	pvs, err := common.RetryOnError(&common.DefaultRetry, common.IsRetriable, func() (interface{}, error) {
 		return client.CoreV1().PersistentVolumes().Get(context.Background(), name, metav1.GetOptions{})
 	})
 	if err != nil {
